app/requirements.txt
Acquisition==5.2
annotated-types==0.6.0
anyio==4.2.0
anykeystore==0.2
astroid==3.0.3
asyncio==3.4.3
attrs==23.2.0
AuthEncoding==5.0
beautifulsoup4==4.12.3
black==24.3.0
bs4==0.0.2
BTrees==5.2
certifi==2024.2.2
cffi==1.16.0
Chameleon==4.5.2
charset-normalizer==3.3.2
click==8.1.7
coverage==6.5.0
coveralls==3.3.1
cryptography==42.0.5
DateTime==5.4
defusedxml==0.7.1
dill==0.3.8
ExtensionClass==5.1
fastapi==0.109.2
flake8==7.0.0
flake8-html==0.4.3
Flask-Login==0.6.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
hupper==1.12.1
idna==3.6
inflect==7.0.0
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
logtail-python==0.2.10
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mirakuru==2.5.2
MultiMapping==5.0
multipart==0.2.4
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
PasteDeploy==3.1.0
pathspec==0.12.1
Persistence==4.1
persistent==5.2
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
pluggy==1.4.0
port-for==0.7.2
psutil==5.9.8
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pyasn1==0.5.1
pycodestyle==2.11.1
pycountry==23.12.11
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-mock==3.14.0
pytest-postgresql==5.1.0
python-dotenv==1.0.1
python-gettext==5.0
python-multipart==0.0.9
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
RestrictedPython==7.0
roman==4.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
text-unidecode==1.3
tomlkit==0.12.3
transaction==4.0
translationstring==1.4
trio==0.24.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
venusian==3.1.0
waitress==3.0.0
Wand==0.6.13
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==3.0.1
WSGIProxy2==0.5.1
zope.interface==6.2
